# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Todo Backend API","description":"A simple todo application backend API","version":"1.0.0"},"paths":{"/api/v1/tasks/":{"post":{"tags":["Tasks","Tasks"],"summary":"Create a new task","description":"Create a new task for the authenticated user. The task will be associated with the user's email extracted from the JWT token.","operationId":"create_task_api_v1_tasks__post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}}},"responses":{"200":{"description":"Task created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"422":{"description":"Validation error - Invalid input data"}}},"get":{"tags":["Tasks","Tasks"],"summary":"List user's tasks","description":"Retrieve a paginated list of tasks for the authenticated user with optional filtering.","operationId":"list_tasks_api_v1_tasks__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"completed","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Completed"}},{"name":"priority","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/PriorityEnum"},{"type":"null"}],"title":"Priority"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date To"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Page Size"}}],"responses":{"200":{"description":"List of tasks retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListResponse"}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["Tasks","Tasks"],"summary":"Get a specific task","description":"Retrieve details of a specific task by its ID. The task must belong to the authenticated user.","operationId":"get_task_api_v1_tasks__task_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}}],"responses":{"200":{"description":"Task retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"404":{"description":"Task not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Tasks","Tasks"],"summary":"Update a task","description":"Update an existing task's details. Only the authenticated user who owns the task can update it.","operationId":"update_task_api_v1_tasks__task_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"200":{"description":"Task updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"404":{"description":"Task not found"},"422":{"description":"Validation error - Invalid input data"}}},"delete":{"tags":["Tasks","Tasks"],"summary":"Delete a task","description":"Permanently delete a task. Only the authenticated user who owns the task can delete it.","operationId":"delete_task_api_v1_tasks__task_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}}],"responses":{"200":{"description":"Task deleted successfully","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"404":{"description":"Task not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/complete":{"patch":{"tags":["Tasks","Tasks"],"summary":"Toggle task completion","description":"Toggle the completion status of a task. Updates the completed_at timestamp accordingly.","operationId":"toggle_task_completion_api_v1_tasks__task_id__complete_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}}],"responses":{"200":{"description":"Task completion status toggled successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"401":{"description":"Unauthorized - Invalid or expired token"},"404":{"description":"Task not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PriorityEnum":{"type":"string","enum":["low","medium","high"],"title":"PriorityEnum"},"TaskCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"anyOf":[{"$ref":"#/components/schemas/PriorityEnum"},{"type":"null"}],"default":"medium"},"due_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"}},"type":"object","required":["title"],"title":"TaskCreate"},"TaskListResponse":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Tasks"},"total_count":{"type":"integer","title":"Total Count"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"}},"type":"object","required":["tasks","total_count","page","page_size"],"title":"TaskListResponse"},"TaskResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_email":{"type":"string","title":"User Email"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"$ref":"#/components/schemas/PriorityEnum"},"is_completed":{"type":"boolean","title":"Is Completed"},"due_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","user_email","title","priority","is_completed","created_at","updated_at"],"title":"TaskResponse"},"TaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"anyOf":[{"$ref":"#/components/schemas/PriorityEnum"},{"type":"null"}]},"due_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"},"is_completed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Completed"}},"type":"object","title":"TaskUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}